LICENSE: Apache License 2.0
I hesitated somewhat between Apache (non-copyleft) and GPLv3 (copyleft)
here.  But eventually, I believe it is best to start with more
permissive licence and ideally leave the *framework* that shall evolve
from this under the permissive licence, as it is more of a technical
exercise than anything innovative and really, Noone Should Have To Do
That Work Again - my personal razor for licence decisions.

If we manage to bring non-trivial algorithms and such into YodaQA,
we may reconsider whether not to relicense the appropriate pipeline
components to GPLv3; anyone can always do that thanks to licence
compatibility.  But there is ample time for that at this point.