A rudimental YodaQAApp with interactive io and AnswerRanker AE
This implements an interactive question reader and answer writer,
type system for FinalAnswerCAS, a dummy AnswerRanker analysis
engine (that just generates an Answer FS straight from the artifact)
and a simple aggregate AE + application wrapper around this all.

In other words, we can now read lines from input and echo them back
to the output. ;-)

This provides the required scaffolding and minimal functionality for
adding the rest of the pipeline and then some real AEs.