New (aggregate) stage QuestionAnalysis + annotator WordTokenizer
The QuestionCAS is now annotated by a custom naive tokenizer
WordTokenizer and these annotations are used by PrimarySearch instead
of it doing its own annotations.

This is set up in such a way that WordTokenizer (and other annotators)
can be reused in other stages as well, e.g. AnswerGenerator.  As a type
repository for annotations produced by these, NLPTypes model is
introduced.