data/eval/: Latest performance measurements
9c54ab2 2014-05-09 PassByClue: Allow si... 15/88/200 7.5%/44.0% avgscore 0.249 avgtime 23.141

So at this point, we can generate a candidate answer in 44% of all
questions of the training set, and pick a right answer as the topmost
final one in 7.5% of all questions of the training set.

Note that the training set is far from perfect; ideal performance may
well be around 75%, since there are many questions that are either
ambiguous (there are many Devil Towers), out-of-date (the head of FBI),
have an answer spelled weird (Nobel Peace Prize in 1992), or I can't
understand them myself (money the U.S. supreme court makes).

Increasing accuracy of candidate answer and final answer generation
is still quite independent; my goal is still increasing the proportion
of candidate answers generated for the time being.  The way forward now
is adding other kinds of searches - body search with title as candidate,
and title-in-clue initial passage.