eval: Extend report by rank information and 5->15 top answers
As we are slowly moving from CandidateAnswer to FinalAnswer precision
optimization, these statistics are pretty crucial.

The tsvout-stats script will continue working as usual on both formats,
but we will add rank statistics extraction later.  The new-format files
have 'ovt' instead of 'out' in their name; future backwards-compatible
but major revisions might be called 'owt' and so on.