Revamp passage scoring: Introduced PassageFeature and PassScoreSimple
Rather than directly assigning the score in PassByClue, a bunch of
features is stored in the CAS and a separate step converts them to
score.  This step will be a machine learning model in the future,
but right now it's just the same equation we used up to now.

(Except that we lose a sqrt(), but that doesn't change the eventual
ordering, obviously.)