AF_OriginMultiple: New feature
Ideally, we would have a multi-level ANN for figuring out these
aggregate features, but before we move on to this classifier, this in
particular seems extremely desirable.

E.g. @originPsgNE is a (surprisingly strong) negative signal, but that's
more likely for the cases where it's a sole origin of the answer and it
shouldn't reduce the score of an answer that was also suggested in other
ways.

This is an even more desirable feature now that the new StanfordParser
version produces much less NP phrases than before.