YodaQA pipeline: Add support for AnswerHitlist (de)serialization
We can now save the AnswerHitlist before final scoring, and restart
the pipeline only for re-scoring.  This means we do not have to
rerun the whole process when we re-train AnswerScore* but just
re-run the scoring.  (The scripts still need to be tweaked for this.)

(While this is pretty nice in itself, this is going to be quite
essential when evidence gathering and second phase scoring is
implemented.  We will have two serialization points then.)