DBpedia: Migrate from Virtuoso to Jena Fuseki as backend database
This is a backwards incompatible change, sorry - if you are using an
older version of YodaQA, you need to update the endpoint.  We still use
DBpedia 3.9 datasets, though.

Virtuoso really sucks.  Its queries are maybe a little faster, but it
can't keep its database in a clean state.  A mere restart of the daemon
seems to be enough to get stuck indefinitely on some queries, and
a clean import of data will result in corrupted database warnings and
the same thing.