AnswerGSHook: Revamp the 'train from the only first correct answer' idea
The problem is that our model training is highly non-stationary now (new
trained model depends on the previous model - this is not desirable) and
it is a real problem both in terms of performance drift and
degeneration.

Therefore, do not do this in the initial phase, making phase 1 the first
natural 'pruning' phrase from the angle of both reduce-to-100 and this,
and in the followup phases use the score from the previous phase for
correct answer selection. This keeps the nifty feature but makes the
model stationary again.