extJWNL Dictionary.lookupIndexWord() -> getIndexWord() except in LATByWordnet
extJWNL, compared to JWNL, unhelpfully tries to specially handle
multi-word queries by looking up their parts.  Somehow, we trigger
a combinatoric explosion and semi-infinite loop by query

	a hot Mediterranean climate (Köppen "Csa"), with subtropical temperatures but a well-defined summer dry season and cool, foggy, rainy season

so we need to switch to a rawer method in LATByWnInstance even though we
lose stemming.

DBpediaWNTypes should be able to switch too since it already contains
a raw word.

LATByWordnet shall not switch yet since the LATs (compared to raw
answers) are not so wildly formed.  But it might still be beneficial
(should we generate hypernymes for "member" when looking at "member
states of NATO"?), needs separate testing.