Add AnswerResource featurestructure tracking answer concepts
This is specific to addressable concept entities generated by solrdoc
and structured answer producers.

This is useful, because:

(i) the dashboard and web interface, first step in an interactive answer
list (and justification tracking might be modelled after this, though
it's apparent that the mergers need a total overhaul.  But as a cop-out
for the short term, we might just add a "kind-of" attribute to
AnswerResource to distinguish the current "conceptIRI" from future
stuff like "originIRI" etc., or even better sub-class the fs.

(ii) the QALD5 challenge requires returning dbpedia resource IRIs very
often, rather than plain strings.  This is the immediate motivation now.

For debugging, this also includes the interactive output printout of
this; other io support may come in the future.