Use *.ailao.eu DNS named for various search endpoints
We may want to start scaling out the databases to the cloud in not so
distant future, and this gives us that flexibility early when we come
back to older YodaQA versions.  Plus, the AAAA record for
freebase.ailao.eu is a tangible immediate benefit as there is a rather
semi-fragile portforwarding scheme in place over IPv4.