Fix UIMA-3818 'Character reference ... is an invalid XML character' error
This would happen during deserialization of XMIs, but the bug is already
in the serialization when a high-codepoint character is serialized in
UTF16 encoding as two XML entities, instead of a single one.  It is
apparently caused by too old XML libraries brought in by Stanford NLP
dependencies, overriding the versions that are part of JDKs, that have
this bug fixed.