DBpediaTitles: Fetch details on label-lookup results from DBpedia again
queryTitleForm() -> queryArticle().

We revert to this to also catch the disambiguations; we were covering
the pageIDs and redirects within the label-lookup service, but this is
no real gain for us anymore and would get qutie unwieldy with the
disambiguation dataset too.

I expect we should be able to go back and greatly simplify label-lookup
by dropping the extra pageIDs, redirect datasets.