CluesToConcepts: clueArticles -> clueBestDists; ignore all subdued clue labels
The new name better expresses the purpose of this map, and the distance
is resolved at the addition rather than at the subduing time, which is
better semantic locality.

Also, this fixes a bug where if multiple ClueLabel instances were
produced for a single subdued label, only the first one was removed.