CluesToConcepts [rewrite]: Group articles by clues, concepts by labels
Our previous implementation had the systematic problem than a clue was
listed in the queue as many times as there were concepts pertaining to
it.  This triggerred various oddities like subduing performed many times
for a single clue, or even a clue added many times for a single label.
These probably weren't effectively bugs, but made the control flow
rather hard to follow and quite fragile with respects to any changes.

This rewrite attempts to rectify this.  Hopefully, the code is also
quite simpler and easier to follow now.