Added Docker Compose file.
Added Docker Compose file that pulls images from Docker Hub. It assumes that the paths actually point to a directory with the necessary files. If this is the case, it can be simply run with docker-compose up. If you remove the k0105/ prefix the same file can be used for locally built images with the Dockerfiles in the respective subdirectories.